moving test models to models/tests into multiple files
